LittleLemon/requirements.txt
django == '*'
mysqlclient == '*'
djangorestframework = '*'